chore(worker): Refactor file paths and module names for entity-related code to feature-related code
