+++ worker/tests/fixture/workflow/feature/filter.yaml
@@ -1,5 +1,5 @@
 id: 7b66c0a4-e1fa-41dd-a0c9-df3f6e01cc22
-name: "entity filter integration test"
+name: "feature filter integration test"
 entryGraphId: c6863b71-953b-4d15-af56-396fc93fc617
 graphs:
   - id: c6863b71-953b-4d15-af56-396fc93fc617
@@ -12,7 +12,7 @@ graphs:
         with:
           format: json
           dataset: |
-            "ram:///fixture/testdata/entity/filter/filter.json"
+            "ram:///fixture/testdata/feature/filter/filter.json"
 
       - id: a1a91180-ab88-4c1a-aab5-48c242a218cb
         name: Ops-StringSpread-Filter
