+++ worker/tests/testcase.rs
@@ -1,3 +1,3 @@
 mod attribute;
-mod entity;
+mod feature;
 mod file;
